Move from traceur to babel
