+++ .eslintrc
@@ -1,0 +1,3 @@
+{
+  "extends": "appium"
+}
