+++ .gitignore
@@ -1,4 +1,3 @@
 node_modules
-lib/es5
-test/es5
 *.log
+build
