+++ lib/main.js
@@ -14,7 +14,7 @@ function mapify (obj) {
     if (obj.hasOwnProperty(k)) {
       m.set(k, mapify(obj[k]));
     }
-  };
+  }
   return m;
 }
 
@@ -33,9 +33,8 @@ function demapify (map) {
     obj[k] = demapify(v);
   }
   return obj;
-};
+}
 
 let objify = demapify;
 
 export { mapify, demapify, objify };
-
